move index file
